OpenDSSC/Pointers and ints: fix sizes (some pointers were being truncated to 32-bit ints)
Also uses `NOMINMAX` to avoid using the `min` and `max` macros from "windows.h"

git-svn-id: https://svn.code.sf.net/p/electricdss/code@3794 d8739450-1e93-4ef4-a0af-c327d92816ff

+++ trunk/VersionC/Support/d2c_sysfile.cpp
@@ -1583,15 +1583,15 @@ void Reset( TTextRec& t )
 }
 
 
-int GetFileHandle( file& F )
+intptr_t GetFileHandle( file& F )
 {
-  return (int) ((TFileRec*) &F )->Handle;
+  return (intptr_t) ((TFileRec*) &F )->Handle;
 }
 
 
-int GetFileHandle( TTextRec& F )
+intptr_t GetFileHandle( TTextRec& F )
 {
-  return (int) ((TTextRec*) &F )->Handle;
+  return (intptr_t) ((TTextRec*) &F )->Handle;
 }
